+++ packages/menu/package.json
@@ -20,7 +20,6 @@
     "access": "public"
   },
   "dependencies": {
-    "@leafygreen-ui/box": "^3.1.1",
     "@leafygreen-ui/emotion": "^4.0.3",
     "@leafygreen-ui/hooks": "^7.3.3",
     "@leafygreen-ui/icon": "^11.12.4",
@@ -28,6 +27,7 @@
     "@leafygreen-ui/lib": "^10.0.0",
     "@leafygreen-ui/palette": "^3.4.7",
     "@leafygreen-ui/popover": "^11.0.4",
+    "@leafygreen-ui/polymorphic": "1.1.0",
     "@leafygreen-ui/tokens": "^2.0.0",
     "lodash": "^4.17.21",
     "react-transition-group": "^4.4.1"
