+++ packages/menu/src/MenuContext/MenuContext.types.ts
@@ -1,0 +1,6 @@
+import { Theme } from '@leafygreen-ui/lib';
+
+export interface MenuData {
+  theme: Theme;
+  darkMode: boolean;
+}
