+++ packages/menu/src/MenuContext/index.ts
@@ -1,0 +1,1 @@
+export { default as MenuContext } from './MenuContext';
