+++ packages/menu/src/MenuGroup/index.ts
@@ -1,0 +1,1 @@
+export { MenuGroup } from './MenuGroup';
