+++ packages/menu/src/MenuItem/MenuItem.styles.ts
@@ -1,0 +1,12 @@
+import { css } from '@leafygreen-ui/emotion';
+import { Theme } from '@leafygreen-ui/lib';
+import { palette } from '@leafygreen-ui/palette';
+
+export const disabledIconStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.gray.dark2};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.gray.base};
+  `,
+};
