+++ packages/menu/src/MenuItem/index.ts
@@ -1,0 +1,6 @@
+export {
+  FocusableMenuItem,
+  FocusableMenuItemElement,
+} from './FocusableMenuItem';
+export { MenuItem, MenuItemElement } from './MenuItem';
+export { MenuItemProps } from './MenuItem.types';
