+++ packages/menu/src/MenuSeparator/index.ts
@@ -1,0 +1,1 @@
+export { MenuSeparator, MenuSeparatorElement } from './MenuSeparator';
