+++ packages/menu/src/SubMenu/index.ts
@@ -1,0 +1,2 @@
+export { SubMenu, SubMenuElement } from './SubMenu';
+export { SubMenuProps } from './SubMenu.types';
